pyproject.toml
[project]
name = "banger-link"
version = "2.0.0"
description = "A Telegram bot that turns any music link into links for every other major streaming service, with chat-level reactions and leaderboards."
authors = [{ name = "Luis Gonçalves", email = "mail@luisg.me" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Chat",
]
dependencies = [
    "python-telegram-bot[job-queue]>=21.6,<23",
    "httpx>=0.27",
    "aiosqlite>=0.20",
    "aiohttp>=3.10",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
]

[project.scripts]
banger-link = "banger_link.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["banger_link"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # handled by formatter
    "RUF001", # ambiguous unicode is intentional in user-facing copy (×, —, …)
    "RUF002",
    "RUF003",
]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
# python-telegram-bot exposes deeply-generic Application; subscribing it
# everywhere is more noise than signal.
disallow_any_generics = false
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--cov=banger_link",
    "--cov-report=term-missing",
    "--cov-report=xml:coverage.xml",
]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"

[tool.coverage.run]
branch = true
source = ["banger_link"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
